Fixed a fatal RecyclerView.Adapter bug
Every time the Feed was loaded after switching to the Resources screen, the app crashed with an IndexOutOfBounds exception. The onBindViewHolder method for the Adapter was passing a position that was out of bounds. Turns out the problem was that onCreateView for the FeedFragment was adding two new FeedItems to the view every time it was called, and it was being called every time the user scrolled to the Resources screen (since the Feed was no longer adjacent to the current Fragment). This was fixed by simply adding a check to see if the FeedItems had already been initialized.